requirements.txt
Django==4.0.4
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
whitenoise==5.2.0
djangorestframework==3.13.1
black
flake8-isort
django-utils-six
python-decouple
djongo
pymongo[srv]==3.12.3